Renamed test packages to reduce imports
