+++ src/test/java/application/data/AmountOfMoneyTest.java
@@ -1,6 +1,5 @@
-package DataTest;
+package application.data;
 
-import application.data.BankAccount;
 import org.junit.jupiter.api.Test;
 
 public class AmountOfMoneyTest {
